+++ scripts/release-plugin/src/snowflakecli_internal_release_plugin/utils.py
